+++ libs/gesha-core/src/error.rs
@@ -3,7 +3,7 @@ pub type Result<A> = std::result::Result<A, Error>;
 #[derive(Debug)]
 pub enum Error {
     IncompatibleVersion,
-    IncompatibleType,
+    TypeMismatch,
     TODO(String),
 }
